+++ frontend/alert/src/components/managealert/AlertItemEnums.js
@@ -1,0 +1,4 @@
+// Enums for AlertItem component
+export const AlertAction = Object.freeze({ Resubscribe: "resub", Cancel: "cancel" });
+export const AlertStatus = Object.freeze({ Closed: "closed", Open: "open" });
+export const AlertRepeat = Object.freeze({ Inactive: "inactive", EOS: "eos", Once: "once" });
